requirements.txt
openfga_sdk
requests
python-dotenv